Do not convert dash to underscore when indexing
Dash (-) was converted to underscore (_) when indexing. This makes it
impossible to search for strings like "non-terminating" (instead one had to
search for "non terminating" which is only an approximation).

In particular, this conversion prevents me from searching for my company's
internal ticket numbers, which contain dashes. Other mail indexers, like
Thunderbird and notmuch, do not have a problem with searching for strings
with dashes.